bruker historikk og tilhørende typer fra amt-lib
