+++ src/main/kotlin/no/nav/amt/deltaker/deltaker/VedtakService.kt
@@ -2,11 +2,11 @@ package no.nav.amt.deltaker.deltaker
 
 import no.nav.amt.deltaker.deltaker.db.VedtakRepository
 import no.nav.amt.deltaker.deltaker.model.Deltaker
-import no.nav.amt.deltaker.deltaker.model.Vedtak
 import no.nav.amt.deltaker.hendelse.HendelseService
 import no.nav.amt.deltaker.hendelse.model.HendelseType
 import no.nav.amt.deltaker.navansatt.NavAnsatt
 import no.nav.amt.deltaker.navansatt.navenhet.NavEnhet
+import no.nav.amt.lib.models.deltaker.Vedtak
 import java.time.LocalDateTime
 import java.util.UUID
