+++ src/main/kotlin/no/nav/amt/deltaker/deltaker/api/model/DeltakelserResponse.kt
@@ -1,6 +1,6 @@
 package no.nav.amt.deltaker.deltaker.api.model
 
-import no.nav.amt.deltaker.deltaker.model.DeltakerStatus
+import no.nav.amt.lib.models.deltaker.DeltakerStatus
 import java.time.LocalDate
 import java.util.UUID
