+++ src/main/kotlin/no/nav/amt/deltaker/deltaker/db/DeltakerEndringRepository.kt
@@ -5,8 +5,8 @@ import kotliquery.Row
 import kotliquery.queryOf
 import no.nav.amt.deltaker.application.plugins.objectMapper
 import no.nav.amt.deltaker.deltaker.forslag.ForslagRepository
-import no.nav.amt.deltaker.deltaker.model.DeltakerEndring
 import no.nav.amt.deltaker.utils.toPGObject
+import no.nav.amt.lib.models.deltaker.DeltakerEndring
 import no.nav.amt.lib.utils.database.Database
 import java.util.UUID
