+++ src/main/kotlin/no/nav/amt/deltaker/deltaker/db/VedtakRepository.kt
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.module.kotlin.readValue
 import kotliquery.Row
 import kotliquery.queryOf
 import no.nav.amt.deltaker.application.plugins.objectMapper
-import no.nav.amt.deltaker.deltaker.model.Vedtak
 import no.nav.amt.deltaker.utils.toPGObject
+import no.nav.amt.lib.models.deltaker.Vedtak
 import no.nav.amt.lib.utils.database.Database
 import java.util.UUID
