+++ src/main/kotlin/no/nav/amt/deltaker/job/DeltakerStatusOppdateringService.kt
@@ -3,8 +3,8 @@ package no.nav.amt.deltaker.job
 import no.nav.amt.deltaker.deltaker.DeltakerService
 import no.nav.amt.deltaker.deltaker.db.DeltakerRepository
 import no.nav.amt.deltaker.deltaker.model.Deltaker
-import no.nav.amt.deltaker.deltaker.model.DeltakerStatus
 import no.nav.amt.deltaker.deltaker.model.harIkkeStartet
+import no.nav.amt.lib.models.deltaker.DeltakerStatus
 import org.slf4j.LoggerFactory
 import java.time.LocalDate
 import java.time.LocalDateTime
